backend/requirements.txt
fastapi
uvicorn
pydantic
llama-cpp-python
huggingface_hub
python-multipart
